functions to utils
