+++ python/rikai/spark/functions/vision.py
@@ -17,7 +17,6 @@
 
 # Standard library
 import os
-import uuid
 from typing import Union
 
 # Third Party
@@ -59,7 +58,6 @@ __all__ = [
     "numpy_to_image",
     "video_to_images",
     "spectrogram_image",
-    "track_detections",
 ]
 
 
@@ -297,103 +295,3 @@ def tracker_match(trackers, detections, bbox_col="bbox", threshold=0.3):
     rows, cols = zip(*np.where(matches))
     idx_map = {cols[idx]: rows[idx] for idx in range(len(rows))}
     return idx_map
-
-
-def match_annotations(iterator, segment_id="vid", id_col="tracker_id"):
-    """
-    Used by mapPartitions to iterate over the small chunks of our hierarchically-organized data.
-    """
-
-    matched_annots = []
-    for idx, data in enumerate(iterator):
-        data = data[1]
-        if not idx:
-            old_row = {idx: uuid.uuid4() for idx in range(len(data[1]))}
-            old_row[segment_id] = data[0]
-            pass
-        annots = []
-        curr_row = {segment_id: data[0]}
-        if old_row[segment_id] != curr_row[segment_id]:
-            old_row = {}
-        if data[2] is not None:
-            for ky, vl in data[2].items():
-                detection = data[1][vl].asDict()
-                detection[id_col] = old_row.get(ky, uuid.uuid4())
-                curr_row[vl] = detection[id_col]
-                annots.append(Row(**detection))
-        matched_annots.append(annots)
-        old_row = curr_row
-    return matched_annots
-
-
-annot_schema = ArrayType(
-    StructType(
-        [
-            StructField("bbox", Box2dType(), False),
-            StructField("tracker_id", StringType(), False),
-        ]
-    )
-)
-
-
-def track_detections(
-    df, segment_id="vid", frames="frame", detections="detections"
-):
-    # utilities
-    id_col = "tracker_id"
-    frame_window = Window().orderBy(frames)
-    value_window = Window().orderBy("value")
-    annot_window = Window.partitionBy(segment_id).orderBy(segment_id, frames)
-    indexer = StringIndexer(inputCol=segment_id, outputCol="vidIndex")
-
-    # dataframe manipulation
-    df = (
-        df.select(segment_id, frames, detections)
-        .withColumn("bbox", F.explode(detections))
-        .withColumn(id_col, F.lit(""))
-        .withColumn("trackables", struct([col("bbox"), col(id_col)]))
-        .groupBy(segment_id, frames, detections)
-        .agg(collect_list("trackables").alias("trackables"))
-        .withColumn(
-            "old_trackables", lag(col("trackables")).over(annot_window)
-        )
-        .withColumn(
-            "matched", tracker_match(col("trackables"), col("old_trackables"))
-        )
-        .withColumn("frame_index", row_number().over(frame_window))
-    )
-
-    # update ids
-    df = (
-        indexer.fit(df)
-        .transform(df)
-        .withColumn("vidIndex", col("vidIndex").cast(StringType()))
-    )
-    unique_ids = df.select("vidIndex").distinct().count()
-    matched = (
-        df.select("vidIndex", segment_id, "trackables", "matched")
-        .rdd.map(lambda x: (x[0], x[1:]))
-        .partitionBy(unique_ids, lambda x: int(x[0]))
-        .mapPartitions(match_annotations)
-    )
-    matched_annotations = spark.createDataFrame(
-        matched, annot_schema
-    ).withColumn("value_index", row_number().over(value_window))
-
-    return (
-        df.join(matched_annotations, col("value_index") == col("frame_index"))
-        .withColumnRenamed("value", "trackers_matched")
-        .withColumn("tracked", explode(col("trackers_matched")))
-        .select(
-            segment_id,
-            frames,
-            detections,
-            col("tracked.{}".format("bbox")).alias("bbox"),
-            col("tracked.{}".format(id_col)).alias(id_col),
-        )
-        .withColumn(id_col, sha2(concat(col(segment_id), col(id_col)), 256))
-        .withColumn("tracked_detections", struct([col("bbox"), col(id_col)]))
-        .groupBy(segment_id, frames, detections)
-        .agg(F.collect_list("tracked_detections").alias("tracked_detections"))
-        .orderBy(segment_id, frames, detections)
-    )
